Moved the Zod types for the AC schema out into their own file, to allow re-use by the preprocessing script.
